+++ src/NextGen/DonationForm/Blocks/DonationFormBlock/Block.php
@@ -2,11 +2,8 @@
 
 namespace Give\NextGen\DonationForm\Blocks\DonationFormBlock;
 
-use Give\Framework\EnqueueScript;
+use Give\NextGen\DonationForm\Actions\GenerateDonationFormViewRouteUrl;
 use Give\NextGen\DonationForm\Blocks\DonationFormBlock\DataTransferObjects\BlockAttributes;
-use Give\NextGen\DonationForm\Repositories\DonationFormRepository;
-use Give\NextGen\DonationForm\ViewModels\DonationFormViewModel;
-use Give\NextGen\Framework\FormTemplates\Registrars\FormTemplateRegistrar;
 
 class Block
 {
@@ -42,95 +39,8 @@ class Block
             return null;
         }
 
-        //$viewUrl = (new GenerateDonationFormViewRouteUrl())($blockAttributes->formId);
+        $viewUrl = (new GenerateDonationFormViewRouteUrl())($blockAttributes->formId, $blockAttributes->formTemplateId);
 
-        //return "<iframe src='$viewUrl'>";
-
-        $viewModel = new DonationFormViewModel($blockAttributes);
-
-        $exports = $viewModel->exports();
-
-        ob_start();
-        ?>
-
-        <script>window.giveNextGenExports = <?= wp_json_encode($exports) ?>;</script>
-
-        <?php
-        $this->enqueueScripts($blockAttributes->formId, $blockAttributes->formTemplateId);
-        ?>
-
-        <div id="root-give-next-gen-donation-form-block"></div>
-
-        <?php
-        return ob_get_clean();
-    }
-
-    /**
-     * Loads scripts in order: [Registrars, Template, Gateways, Block]
-     *
-     * @unreleased
-     *
-     * @return void
-     */
-    private function enqueueScripts(int $formId, string $formTemplateId)
-    {
-        /** @var DonationFormRepository $donationFormRepository */
-        $donationFormRepository = give(DonationFormRepository::class);
-
-        // load registrars
-        (new EnqueueScript(
-            'give-donation-form-registrars-js',
-            'build/donationFormRegistrars.js',
-            GIVE_NEXT_GEN_DIR,
-            GIVE_NEXT_GEN_URL,
-            'give'
-        ))->loadInFooter()->enqueue();
-
-        // load template
-        /** @var FormTemplateRegistrar $formTemplateRegistrar */
-        $formTemplateRegistrar = give(FormTemplateRegistrar::class);
-
-        // silently fail if template is missing for some reason
-        if ($formTemplateRegistrar->hasTemplate($formTemplateId)) {
-            $template = $formTemplateRegistrar->getTemplate($formTemplateId);
-
-            if ($template->css()) {
-                wp_enqueue_style('givewp-form-template-' . $template->getId(), $template->css());
-            }
-
-            if ($template->js()) {
-                wp_enqueue_script(
-                    'givewp-form-template-' . $template->getId(),
-                    $template->js(),
-                    array_merge(
-                        ['give-donation-form-registrars-js'],
-                        $template->dependencies()
-                    ),
-                    false,
-                    true
-                );
-            }
-        }
-
-        // load gateways
-        foreach ($donationFormRepository->getEnabledPaymentGateways($formId) as $gateway) {
-            if (method_exists($gateway, 'enqueueScript')) {
-                /** @var EnqueueScript $script */
-                $script = $gateway->enqueueScript();
-
-                $script->dependencies(['give-donation-form-registrars-js'])
-                    ->loadInFooter()
-                    ->enqueue();
-            }
-        }
-
-        // load block - since this is using render_callback viewScript in blocks.json will not work.
-        (new EnqueueScript(
-            'give-next-gen-donation-form-block-js',
-            'build/donationFormBlockApp.js',
-            GIVE_NEXT_GEN_DIR,
-            GIVE_NEXT_GEN_URL,
-            'give'
-        ))->dependencies(['give-donation-form-registrars-js'])->loadInFooter()->enqueue();
+        return "<div style='position: relative;width: 100%;height: 100%;overflow: hidden;margin: 0 auto;padding-top: 100%;'><iframe src='$viewUrl' style='position: absolute;top: 0;left: 0;bottom: 0;right: 0;width: 100%;height: 100%;'></div>";
     }
 }
